Bug-fix: Fixed syntax error in StorageRef, Also loading bug in file uplaod
The storageRef line was having a syntax error, firebase.storage() is a method.
The loading spinner was not stopping even after successful upload of the image.
Added a conditional to indicate the completion of the image upload to firebase storage.
Also the 'isUplaoding' was not getting false on error, setting it to false if error occures
.